requirements.txt
fastapi
uvicorn
httpx
pydantic>=2.0
pydantic-settings
pytest
pytest-asyncio
python-dotenv
jinja2
